+++ test_umbra/test_experiment.py
@@ -7,16 +7,13 @@ from umbra import experiment
 from .test_common import TestBase
 
 class TestLoadMetadata(TestBase):
-    """Test basic metadata loading.
+    """Test basic metadata loading."""
 
-    Each version of metadata.csv should provide the same list of
-    dictionaries.
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.exp_path = self.paths["exp"] / "Partials_1_1_18"
-        self.expected = [ \
+    def test_load_metadata(self):
+        """Test basic metadata loading."""
+        fp_metadata = self.path / "metadata.csv"
+        info = experiment.load_metadata(fp_metadata)
+        expected = [ \
             OrderedDict([
                 ('Sample_Name', '1086S1_01'), ('Project', 'STR'),
                 ('Contacts', {'Jesse': 'ancon@upenn.edu'}), ('Tasks', ['trim'])]),
@@ -30,33 +27,52 @@ class TestLoadMetadata(TestBase):
                 ('Sample_Name', '1086S1_04'), ('Project', 'Something Else'),
                 ('Contacts', {'Someone': 'person@gmail.com', 'Jesse Connell': 'ancon@upenn.edu'}),
                 ('Tasks', [])])]
+        self.assertEqual(expected, info)
+
+
+class TestLoadMetadataCaps(TestLoadMetadata):
+    """Test metadata loading with capital letters in the Tasks column."""
+
+
+class TestLoadMetadataEmptyrows(TestLoadMetadata):
+    """Test metadata loading including empty rows.
+
+    Empty rows should be silently removed.
+    """
+
+
+class TestLoadMetadataEmptycols(TestLoadMetadata):
+    """Test metadata loading including empty columns.
+
+    Empty columns should be silently removed.
+    """
+
+
+class TestLoadMetadataISO8859(TestLoadMetadata):
+    """Test metadata loading with non-unicode file.
+
+    By default the file parsing should fail and raise an exception, but we can
+    override the behavior.
+    """
 
     def test_load_metadata(self):
         """Test basic metadata loading."""
-        fp_metadata = self.exp_path / "metadata.csv"
-        info = experiment.load_metadata(fp_metadata)
-        self.assertEqual(self.expected, info)
-
-    def test_load_metadata_caps(self):
-        """Test metadata loading with capital letters in the Tasks column."""
-        fp_metadata = self.exp_path / "metadata_caps.csv"
-        info = experiment.load_metadata(fp_metadata)
-        self.assertEqual(self.expected, info)
-
-    def test_load_metadata_emptyrows(self):
-        """Test metadata loading including empty rows.
-
-        Empty rows should be silently removed.
-        """
-        fp_metadata = self.exp_path / "metadata_emptyrows.csv"
-        info = experiment.load_metadata(fp_metadata)
-        self.assertEqual(self.expected, info)
-
-    def test_load_metadata_emptycols(self):
-        """Test metadata loading including empty columns.
-
-        Empty columns should be silently removed.
-        """
-        fp_metadata = self.exp_path / "metadata_emptycols.csv"
-        info = experiment.load_metadata(fp_metadata)
-        self.assertEqual(self.expected, info)
+        fp_metadata = self.path / "metadata.csv"
+        with self.assertRaises(UnicodeDecodeError):
+            info = experiment.load_metadata(fp_metadata)
+        info = experiment.load_metadata(fp_metadata, non_unicode="strip")
+        expected = [ \
+            OrderedDict([
+                ('Sample_Name', '1086S1_01'), ('Project', 'STR'),
+                ('Contacts', {'Jesse': 'ancon@upenn.edu'}), ('Tasks', ['trim'])]),
+            OrderedDict([
+                ('Sample_Name', '1086S1_02'), ('Project', 'STR'),
+                ('Contacts', {'Jesse': 'ancon@upenn.edu'}), ('Tasks', ['trim'])]),
+            OrderedDict([
+                ('Sample_Name', '1086S1_03'), ('Project', 'Something Else'),
+                ('Contacts', {'Someone': 'person@gmail.com'}), ('Tasks', [])]),
+            OrderedDict([
+                ('Sample_Name', '1086S1_04'), ('Project', 'Something Else'),
+                ('Contacts', {'Someone': 'person@gmail.com', 'Jesse Connell': 'ancon@upenn.edu'}),
+                ('Tasks', [])])]
+        self.assertEqual(expected, info)