+++ test_umbra/test_illumina/test_run.py
@@ -5,190 +5,323 @@ More specifically, test the Run class that represents an Illumina run directory
 on disk.
 """
 
-import unittest
 from unittest.mock import Mock
 import time
 import datetime
 import warnings
-from tempfile import TemporaryDirectory
-from pathlib import Path
-from distutils.dir_util import copy_tree
-from shutil import move
 from umbra.illumina.run import Run
-from .test_common import RUN_IDS, PATH_RUNS
+from .test_common import dummy_bcl_stats
+from ..test_common import TestBase
 
-class TestRun(unittest.TestCase):
-    """Base test case for a Run.
 
-    This is built around a mock MiSeq run directory.
+class TestRun(TestBase):
+    """Basic test case for a typical Run."""
+
+    def setUp(self):
+        super().setUp()
+        self.aln_callback = Mock()
+        self.run = Run(
+            self.path / "180101_M00000_0000_000000000-XXXXX",
+            alignment_callback=self.aln_callback)
+
+    def test_rta_complete(self):
+        """Test the rta_complete property.
+
+        This should be a small dictionary parsed parsed from RTAComplete.txt.
+        """
+        self.assertEqual(
+            self.run.rta_complete,
+            {
+                "Date": datetime.datetime(2018, 1, 1, 6, 21, 31, 705000),
+                "Version": "Illumina RTA 1.18.54"})
+
+    def test_completed_job_info(self):
+        """Test the completed_job_info property.
+
+        This shuld be a data structure parsed from CompletedJobInfo.xml.
+        """
+        self.assertEqual(
+            self.run.completed_job_info.find("StartTime").text,
+            "2018-01-02T06:48:22.0480092-04:00")
+        self.assertEqual(
+            self.run.completed_job_info.find("CompletionTime").text,
+            "2018-01-02T06:48:32.608024-04:00")
+
+    def test_refresh(self):
+        """Test the refresh method.
+
+        This should refresh the basic run data from disk, load any new
+        alignments, and refresh any existing alignments.
+
+        For a complete run, refresh has no effect.
+        """
+        self.aln_callback.assert_called_once()
+        self.run.refresh()
+        self.aln_callback.assert_called_once()
+
+    def test_load_all_bcl_stats(self):
+        """Test the load_all_bcl_stats method.
+
+        This should load a list of dictionaries, one for each cycle, from the
+        .stats files.
+        """
+        observed = self.run.load_all_bcl_stats()
+        expected = dummy_bcl_stats(16, 1)
+        self.assertEqual(observed, expected)
+
+    def test_run_id(self):
+        """Test the run_id str property.
+
+        This should be the run identifier, usually identical to the directory
+        name.
+        """
+        self.assertEqual(
+            self.run.run_id,
+            "180101_M00000_0000_000000000-XXXXX")
+
+    def test_run_info(self):
+        """Test the run_info XML object property.
+
+        This shuld be a data structure parsed from RunInfo.xml.
+        """
+        self.assertEqual(
+            self.run.run_info.find("Run").attrib["Id"],
+            "180101_M00000_0000_000000000-XXXXX")
+
+    def test_flowcell(self):
+        """Test the flowcell str property.
+
+        This should be the flowcell identifier as provided by the run metadata.
+        """
+        self.assertEqual("000000000-XXXXX", self.run.flowcell)
+
+    def test_complete(self):
+        """Test the complete property (True for a completed run)."""
+        self.assertTrue(self.run.complete)
+
+
+class TestRunIncomplete(TestRun):
+    """Test behavior on an incomplete MiSeq run.
+
+    For the most part the run information will be available but not the
+    rta_complete information.
     """
 
     def setUp(self):
-        self.tmpdir = TemporaryDirectory()
-        self.path_run = Path(self.tmpdir.name) / RUN_IDS["MiSeq"]
-        copy_tree(str(PATH_RUNS / RUN_IDS["MiSeq"]), str(self.path_run))
-        self.run = Run(self.path_run)
-        date = datetime.datetime(2018, 1, 1, 6, 21, 31, 705000)
-        self.expected = {
-            "id": RUN_IDS["MiSeq"],
-            "cycles": 318,
-            "rta": {"Date": date, "Version": "Illumina RTA 1.18.54"},
-            "t1": "2018-01-02T06:48:22.0480092-04:00",
-            "t2": "2018-01-02T06:48:32.608024-04:00",
-            "flowcell": "000000000-XXXXX"
-            }
+        super().setUp()
+        self.aln_callback = Mock()
+        self.run = Run(
+            self.path / "180101_M00000_0000_000000000-XXXXX",
+            alignment_callback=self.aln_callback)
 
-    def tearDown(self):
-        self.tmpdir.cleanup()
+    def test_rta_complete(self):
+        """Test the rta_complete property.
 
-    def test_attrs(self):
-        """Test various Run properties."""
-        # RunInfo.xml
-        # Check the Run ID.
-        id_obs = self.run.run_info.find("Run").attrib["Id"]
-        self.assertEqual(id_obs, self.expected["id"])
-        # RTAComplete.txt
-        # Check the full contents.
-        rta_obs = self.run.rta_complete
-        self.assertEqual(rta_obs, self.expected["rta"])
-        # CompletedJobInfo.xml
-        # Check the job start/completion timestamps.
-        t1_obs = self.run.completed_job_info.find("StartTime").text
-        t2_obs = self.run.completed_job_info.find("CompletionTime").text
-        self.assertEqual(self.expected["t1"], t1_obs)
-        self.assertEqual(self.expected["t2"], t2_obs)
-
-    def check_refresh_run(self):
-        """Test refreshing run state from files on disk."""
-        # Starting without a RTAComplete.txt, run is marked incomplete.  We'll
-        # also check how the alignment refresh behaves.  It shouldn't call the
-        # given callback, if any, until the run itself is complete.  This is an
-        # edge case since in theory the run needs to be complete if the
-        # alignment is, but I did run into this in a real run directory that
-        # was incompletely transferred and this behavior makes the most sense
-        # in that situation.
-        move(str(self.path_run / "RTAComplete.txt"), str(self.path_run / "tmp.txt"))
-        callback = Mock()
-        self.run = Run(self.path_run, alignment_callback=callback)
-        self.assertFalse(self.run.complete)
+        For an incomplete run tis doesn't yet exist.
+        """
         self.assertEqual(self.run.rta_complete, None)
-        callback.assert_not_called()
-        # It doesn't update automatically.
-        move(str(self.path_run / "tmp.txt"), str(self.path_run / "RTAComplete.txt"))
-        self.assertFalse(self.run.complete)
-        # On refresh, it is now seen as complete.
-        self.run.refresh()
-        callback.assert_called_once()
-        self.assertTrue(self.run.complete)
-        self.run.refresh()
-        callback.assert_called_once()
 
-    def check_refresh_alignments(self):
-        """Test refreshing run Alignment directories from files on disk."""
-        orig_als = self.run.alignments
-        path_checkpoint = self.run.alignments[0].paths["checkpoint"]
-        move(str(path_checkpoint), str(self.path_run / "tmp.txt"))
-        self.run = Run(self.path_run)
-        self.assertEqual(len(self.run.alignments), len(orig_als))
-        self.assertFalse(self.run.alignments[0].complete)
-        move(str(self.path_run / "tmp.txt"), str(path_checkpoint))
-        self.assertFalse(self.run.alignments[0].complete)
-        self.run.refresh()
-        self.assertTrue(self.run.alignments[0].complete)
-        # Now, load any new alignments
-        path_al = self.run.alignments[0].path
-        move(str(path_al), str(self.path_run / "tmp"))
-        self.run = Run(self.path_run)
-        self.assertEqual(len(self.run.alignments), len(orig_als)-1)
-        move(str(self.path_run / "tmp"), str(path_al))
-        self.assertEqual(len(self.run.alignments), len(orig_als)-1)
-        self.run.refresh()
-        self.assertEqual(len(self.run.alignments), len(orig_als))
+    def test_complete(self):
+        """Test that the run is not reported as complete."""
+        self.assertFalse(self.run.complete)
 
     def test_refresh(self):
-        """Test that a run refresh works"""
-        self.check_refresh_run() # 1: Update run completion status
-        self.check_refresh_alignments() # 2: refresh existing alignments
+        """For an incomplete run the alignent callback should not be called."""
+        self.aln_callback.assert_not_called()
+        self.run.refresh()
+        self.aln_callback.assert_not_called()
 
-    def test_load_all_bcl_stats(self):
-        """Test loading all of the .stats files into a list."""
-        observed = self.run.load_all_bcl_stats()
-        expected = []
-        for lane in [1]:
-            for cycle in range(self.expected["cycles"]):
-                for tile in [1101, 1102]:
-                    expected.append({
-                        'cycle': cycle,
-                        'avg_intensity': 0.0,
-                        'avg_int_all_A': 0.0,
-                        'avg_int_all_C': 0.0,
-                        'avg_int_all_G': 0.0,
-                        'avg_int_all_T': 0.0,
-                        'avg_int_cluster_A': 0.0,
-                        'avg_int_cluster_C': 0.0,
-                        'avg_int_cluster_G': 0.0,
-                        'avg_int_cluster_T': 0.0,
-                        'num_clust_call_A': 0,
-                        'num_clust_call_C': 0,
-                        'num_clust_call_G': 0,
-                        'num_clust_call_T': 0,
-                        'num_clust_call_X': 0,
-                        'num_clust_int_A': 0,
-                        'num_clust_int_C': 0,
-                        'num_clust_int_G': 0,
-                        'num_clust_int_T': 0,
-                        'lane': lane,
-                        'tile': tile})
-            self.assertEqual(observed, expected)
+class TestRunToComplete(TestBase):
+    """Tests for the incomplete -> complete transition.
 
-    def test_run_id(self):
-        """Test the run ID property."""
-        self.assertEqual(self.expected["id"], self.run.run_id)
+    Instead of a typical TestRun with all methods and properties tested,
+    this checks the behavior of a few key things during the transition from an
+    incomplete state to complete.
+    """
 
-    def test_flowcell(self):
-        """Test the flowcell property."""
-        self.assertEqual(self.expected["flowcell"], self.run.flowcell)
+    def setUp(self):
+        super().setUp()
+        self.aln_callback = Mock()
+        self.run = Run(
+            self.path / "180101_M00000_0000_000000000-XXXXX",
+            alignment_callback=self.aln_callback)
+
+    def tearDown(self):
+        self.reset_complete()
+
+    def make_complete(self):
+        """Create a RTAComplete.txt file so the alignment is complete"""
+        rta = self.run.path / "RTAComplete.txt"
+        with open(rta, "wt") as f_out:
+            f_out.write("1/1/2018,06:21:31.705,Illumina RTA 1.18.54\r\n")
+
+    def reset_complete(self):
+        """Remove the RTAComplete.txt file, if any."""
+        rta = self.run.path / "RTAComplete.txt"
+        try:
+            rta.unlink()
+        except FileNotFoundError:
+            pass
+
+    def test_rta_complete(self):
+        """Test the rta_complete property.
+
+        This should be a small dictionary parsed parsed from RTAComplete.txt.
+        """
+        self.assertIsNone(self.run.rta_complete)
+        self.make_complete()
+        self.assertIsNone(self.run.rta_complete)
+        self.run.refresh()
+        self.assertEqual(
+            self.run.rta_complete,
+            {
+                "Date": datetime.datetime(2018, 1, 1, 6, 21, 31, 705000),
+                "Version": "Illumina RTA 1.18.54"})
+
+    def test_refresh(self):
+        """Does run refresh catch completion?
+
+        At first the alignment callback function should not have been called at
+        all, whatever state of the alignment.  Only when both the run and
+        alignment are complete *and* refresh is called should the callback be
+        called.
+        """
+        self.aln_callback.assert_not_called()
+        self.run.refresh()
+        self.aln_callback.assert_not_called()
+        self.make_complete()
+        self.aln_callback.assert_not_called()
+        self.run.refresh()
+        self.aln_callback.assert_called_once()
+        self.run.refresh()
+        self.aln_callback.assert_called_once()
+
+    def test_complete(self):
+        """Is the Run complete?
+
+        Not until the RTAComplete.txt file arrives and the object is refreshed.
+        """
+        self.assertFalse(self.run.complete)
+        self.make_complete()
+        self.assertFalse(self.run.complete)
+        self.run.refresh()
+        self.assertTrue(self.run.complete)
 
 
-class TestRunSingle(TestRun):
+class TestRunSingle(TestBase):
     """Test Run with single-ended sequencing."""
 
     def setUp(self):
-        self.tmpdir = TemporaryDirectory()
-        self.path_run = Path(self.tmpdir.name) / RUN_IDS["Single"]
-        copy_tree(str(PATH_RUNS / RUN_IDS["Single"]), str(self.path_run))
-        self.run = Run(self.path_run)
-        self.expected = {}
-        self.expected["id"] = RUN_IDS["Single"]
-        self.expected["cycles"] = 518
-        date = datetime.datetime(2018, 1, 6, 6, 20, 25, 841000)
-        self.expected["rta"] = {"Date": date, "Version": "Illumina RTA 1.18.54"}
-        self.expected["t1"] = "2018-01-05T13:38:15.2566992-04:00"
-        self.expected["t2"] = "2018-01-05T13:38:45.3021522-04:00"
-        self.expected["flowcell"] = "000000000-XXXXX"
+        super().setUp()
+        self.aln_callback = Mock()
+        self.run = Run(
+            self.path / "180105_M00000_0000_000000000-XXXXX",
+            alignment_callback=self.aln_callback)
+
+    def test_run_id(self):
+        """Test the run_id str property."""
+        self.assertEqual(
+            self.run.run_id,
+            "180105_M00000_0000_000000000-XXXXX")
+
+    def test_run_info(self):
+        """Test the run_info XML object property."""
+        self.assertEqual(
+            self.run.run_info.find("Run").attrib["Id"],
+            "180105_M00000_0000_000000000-XXXXX")
+
+    def test_flowcell(self):
+        """Test the flowcell str property."""
+        self.assertEqual("000000000-XXXXX", self.run.flowcell)
+
+    def test_rta_complete(self):
+        """Test the rta_complete property."""
+        self.assertEqual(
+            self.run.rta_complete,
+            {
+                "Date": datetime.datetime(2018, 1, 6, 6, 20, 25, 841000),
+                "Version": "Illumina RTA 1.18.54"})
+
+    def test_completed_job_info(self):
+        """Test the completed_job_info property."""
+        self.assertEqual(
+            self.run.completed_job_info.find("StartTime").text,
+            "2018-01-05T13:38:15.2566992-04:00")
+        self.assertEqual(
+            self.run.completed_job_info.find("CompletionTime").text,
+            "2018-01-05T13:38:45.3021522-04:00")
+
+    def test_refresh(self):
+        """Test the refresh method."""
+        self.aln_callback.assert_called_once()
+        self.run.refresh()
+        self.aln_callback.assert_called_once()
+
+    def test_load_all_bcl_stats(self):
+        """Test the load_all_bcl_stats method.
+
+        There are just fewer cycles here with no R2 but otherwise the idea is
+        the same as for the paired-end run.
+        """
+        observed = self.run.load_all_bcl_stats()
+        expected = dummy_bcl_stats(12, 1)
+        self.assertEqual(observed, expected)
 
 
-class TestRunMiniSeq(TestRun):
-    """Like TestRun, but for a MiniSeq run.
+class TestRunMiniSeq(TestBase):
+    """Test MiniSeq run.
 
     Nothing should really change in how the information is presented, just a
     few details that are different for this particular run.  The class should
     abstract away the actual differences in run directory layout between the
-    different sequencer model.s
+    different sequencer models.
     """
 
     def setUp(self):
-        self.tmpdir = TemporaryDirectory()
-        self.path_run = Path(self.tmpdir.name) / RUN_IDS["MiniSeq"]
-        copy_tree(str(PATH_RUNS / RUN_IDS["MiniSeq"]), str(self.path_run))
-        self.run = Run(self.path_run)
-        self.expected = {}
-        self.expected["id"] = RUN_IDS["MiniSeq"]
-        date = datetime.datetime(2018, 1, 4, 11, 14, 00)
-        self.expected["rta"] = {"Date": date, "Version": "RTA 2.8.6"}
-        self.expected["t1"] = "2018-01-04T11:15:03.8237582-04:00"
-        self.expected["t2"] = "2018-08-04T11:16:52.4989741-04:00"
-        self.expected["flowcell"] = "000000000"
+        super().setUp()
+        self.aln_callback = Mock()
+        self.run = Run(
+            self.path / "180103_M000000_0000_0000000000",
+            alignment_callback=self.aln_callback)
+
+    def test_rta_complete(self):
+        """Test the rta_complete property."""
+        self.assertEqual(
+            self.run.rta_complete,
+            {
+                "Date": datetime.datetime(2018, 1, 4, 11, 14, 00),
+                "Version": "RTA 2.8.6"})
+
+    def test_completed_job_info(self):
+        """Test the completed_job_info property."""
+        self.assertEqual(
+            self.run.completed_job_info.find("StartTime").text,
+            "2018-01-04T11:15:03.8237582-04:00")
+        self.assertEqual(
+            self.run.completed_job_info.find("CompletionTime").text,
+            "2018-08-04T11:16:52.4989741-04:00")
+
+    def test_refresh(self):
+        """Test that for a complete run, refresh has no effect."""
+        self.aln_callback.assert_called_once()
+        self.run.refresh()
+        self.aln_callback.assert_called_once()
+
+    def test_run_id(self):
+        """Test the run_id str property."""
+        self.assertEqual(
+            self.run.run_id,
+            "180103_M000000_0000_0000000000")
+
+    def test_run_info(self):
+        """Test the run_info XML object property."""
+        self.assertEqual(
+            self.run.run_info.find("Run").attrib["Id"],
+            "180103_M000000_0000_0000000000")
+
+    def test_flowcell(self):
+        """Test the flowcell str property."""
+        self.assertEqual("000000000", self.run.flowcell)
 
     def test_load_all_bcl_stats(self):
         """Test that a MiniSeq run gives an empty list as it has no BCL stats files."""
@@ -200,18 +333,10 @@ class TestRunMisnamed(TestRun):
     """Test case for a directory whose name is not the Run ID."""
 
     def setUp(self):
-        self.tmpdir = TemporaryDirectory()
-        self.path_run = Path(self.tmpdir.name) / RUN_IDS["misnamed"]
-        copy_tree(str(PATH_RUNS / RUN_IDS["MiSeq"]), str(self.path_run))
-        self.run = Run(self.path_run, strict=False)
-        self.expected = {}
-        self.expected["id"] = RUN_IDS["MiSeq"]
-        self.expected["cycles"] = 318
-        date = datetime.datetime(2018, 1, 1, 6, 21, 31, 705000)
-        self.expected["rta"] = {"Date": date, "Version": "Illumina RTA 1.18.54"}
-        self.expected["t1"] = "2018-01-02T06:48:22.0480092-04:00"
-        self.expected["t2"] = "2018-01-02T06:48:32.608024-04:00"
-        self.expected["flowcell"] = "000000000-XXXXX"
+        TestBase.setUp(self)
+        self.path_run = self.path / "run-files-custom-name"
+        self.aln_callback = Mock()
+        self.run = Run(self.path_run, alignment_callback=self.aln_callback)
 
     def test_init(self):
         """Test that we get the expected warning during Run initialization."""
@@ -229,12 +354,12 @@ class TestRunMisnamed(TestRun):
             self.assertEqual(0, len(warn_list))
 
 
-class TestRunInvalid(unittest.TestCase):
+class TestRunInvalid(TestBase):
     """Test case for a directory that is not an Illumina run."""
 
     def test_init(self):
         """Test that Run initialization handles invalid an run directory."""
-        path = PATH_RUNS / RUN_IDS["not a run"]
+        path = self.path / "not a run"
         with self.assertRaises(ValueError):
             Run(path)
         with warnings.catch_warnings(record=True) as warn_list:
@@ -242,7 +367,7 @@ class TestRunInvalid(unittest.TestCase):
             run = Run(path, strict=False)
             self.assertEqual(1, len(warn_list))
             self.assertTrue(run.invalid)
-        path = PATH_RUNS / RUN_IDS["nonexistent"]
+        path = self.path / "nonexistent"
         with self.assertRaises(ValueError):
             Run(path)
         with warnings.catch_warnings(record=True) as warn_list:
@@ -252,31 +377,33 @@ class TestRunInvalid(unittest.TestCase):
             self.assertTrue(run.invalid)
 
 
-class TestRunMinAlignmentAge(TestRun):
+class TestRunMinAlignmentAge(TestBase):
     """Test case for a Run set to ignore too-new alignment directories.
 
     This should not warn about empty alignment directories if they're newer (by
     ctime) than a certain age."""
 
-    def check_refresh_alignments(self):
-        # We just need to override the alignment refresh check method.  Last
-        # time we didn't have Checkpoint.txt, so it didn't consider it
-        # complete.  This time we'll see what happens with Alignment
-        # directories with no SampleSheetUsed.csv.
+    def setUp(self):
+        super().setUp()
+        self.path_run = self.path / "180101_M00000_0000_000000000-XXXXX"
+        self.run = Run(self.path_run)
+
+    def test_init(self):
+        """Does Run instantiation respect min_alignment_dir_age?"""
         orig_als = self.run.alignments
         # min_alignment_dir_age should make the Run object skip the "too new"
         # alignment directories.  We'll use a one-second setting for this test,
         # and will touch the directory to reset the ctime.
         self.run.alignments[0].path.touch()
         with self.assertLogs(level="DEBUG") as log_context:
-            self.run = Run(self.path_run, min_alignment_dir_age=1)
+            self.run = Run(self.path_run, min_alignment_dir_age=0.2)
         # That alignment shouldn't have been added to the list yet.  No
         # warnings should have been generated, just a debug log message about
         # the skip.
         self.assertEqual(len(log_context.output), 1)
         self.assertEqual(len(self.run.alignments), len(orig_als)-1)
         # After enough time has passed it should be loaded.
-        time.sleep(1.1)
+        time.sleep(0.5)
         self.assertEqual(len(self.run.alignments), len(orig_als)-1)
         self.run.refresh()
         self.assertEqual(len(self.run.alignments), len(orig_als))