+++ CHANGELOG.md
@@ -7,12 +7,20 @@
  * `error` property for `illumina.alignment.Alignment` class ([#104])
  * BCL stats file parsing functions for applicable Illumina runs ([#106])
 
+### Changed
+
+ * Use faster re-tries for small file uploads to Box ([#111])
+ * Use fully-resolved paths for alignment directory tracking in Run.refresh
+   ([#111])
+ * Reorganize unit tests for speed and simplicity ([#111])
+
 ### Fixed
 
  * Only process an alignment when its containing run is marked complete
    ([#108])
  * Handle non-unicode characters in CSV files ([#105])
 
+[#111]: https://github.com/ShawHahnLab/umbra/pull/111
 [#108]: https://github.com/ShawHahnLab/umbra/pull/108
 [#106]: https://github.com/ShawHahnLab/umbra/pull/106
 [#105]: https://github.com/ShawHahnLab/umbra/pull/105